fix build error /home/runner/work/data_station/data_station/LART_Car_Dashboard_v1/src/ui/test_can_bridge.cpp:75:33: warning: comparison is always true due to limited range of data type [-Wtype-limits]

+++ LART_Car_Dashboard_v1/src/ui/test_can_bridge.cpp
@@ -72,7 +72,7 @@ int main(int argc, char **argv) {
     auto response_sub = node->create_subscription<lart_msgs::msg::IcdResponse>(
         "/can/dbc/icd_response", rclcpp::QoS(10).best_effort(),
         [&](lart_msgs::msg::IcdResponse::SharedPtr msg) {
-            if (msg->resp_muxid != 228 || msg->resp_articlenumber != article || msg->resp_fwmajor != 0)
+            if (static_cast<uint8_t>(msg->resp_muxid) != 228 || msg->resp_articlenumber != article || msg->resp_fwmajor != 0)
                 throw std::runtime_error("ICD precision/multiplexer mismatch");
             got_response = true;
         });
